ris/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
Babel==2.14.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.61.1
django-extra-views==0.14.0
django-filter==23.5
django-haystack==3.2.1
django-ipware==6.0.4
django-modelcluster==6.3
django-oscar==3.2.4
django-permissionedforms==0.1
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-taggit==5.0.1
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
draftjs-exporter==5.0.0
drf-recaptcha==3.0.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==23.3.0
filetype==1.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
laces==0.1.1
oauthlib==3.2.2
openpyxl==3.1.2
phonenumbers==8.13.31
pillow==10.2.0
pillow_heif==0.15.0
purl==1.6
pycountry==23.12.11
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-ipware==2.0.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tzdata==2024.1
urllib3==2.2.1
wagtail==6.0.1
webencodings==0.5.1
Willow==1.8.0
